+++ src/debugger/client/jsonlib.d.ts
@@ -1,0 +1,7 @@
+/**
+ * @noResolution
+ */
+declare module "json" {
+    function encode(table: object): string;
+    function decode<T>(json: string): T;
+}