+++ src/debugger/utils.d.ts
@@ -1,0 +1,3 @@
+type DiscriminateUnion<T, K extends keyof T, V extends T[K]> = T extends Record<K, V>
+    ? T
+    : never;