+++ src/extension.ts
@@ -1,14 +1,20 @@
-import * as vscode from "vscode";
-import { init, mount } from "./commands";
+import { init, mount, installClient } from "./commands";
+import { stopListening, startListening } from "./debugger/server/listener";
+import { commands, ExtensionContext, debug } from "vscode";
+import { LuaConfigurationProvider } from "./debugger/server/luaConfigurationProvider";
+import { LuaDebugAdapterFactory } from "./debugger/server/luaDebugAdapterFactory";
 
-export function activate(context: vscode.ExtensionContext): void {
+export function activate(context: ExtensionContext): void {
     context.subscriptions.push(
-        vscode.commands.registerCommand("oc-ts.init", init(context)),
-        vscode.commands.registerCommand("oc-ts.mount", mount())
+        commands.registerCommand("oc-ts.init", init(context)),
+        commands.registerCommand("oc-ts.mount", mount()),
+        commands.registerCommand("oc-ts.installClient", installClient()),
+        debug.registerDebugConfigurationProvider("tsdbg", new LuaConfigurationProvider()),
+        debug.registerDebugAdapterDescriptorFactory("tsdbg", new LuaDebugAdapterFactory())
     );
+    startListening();
 }
 
-// this method is called when your extension is deactivated
 export function deactivate(): void {
-    //
+    stopListening();
 }