+++ src/navigation/navigationItem.ts
@@ -1,11 +1,7 @@
 export class NavigationItem {
-    readonly name?: string;
-    readonly path: string;
-    readonly virtualPath: string;
-
-    constructor(path: string, virtualPath: string, name?: string) {
-        this.path = path;
-        this.virtualPath = virtualPath;
-        this.name = name;
-    }
+    public constructor(
+        public readonly path: string,
+        public readonly virtualPath: string,
+        public readonly name?: string
+    ) {}
 }